fixed error
An error was displayed that React could not output a list of objects. The error was that React was trying to render objects in the SheduleCard component that it did not receive from anywhere. Because of this, an error was displayed. The bug has been fixed, at the moment the component does not output anything, since redux and context have not yet been completed